+++ apps/dashboard/src/app/components/label/label-field/label-field.component.ts
@@ -1,5 +1,5 @@
 import { ChangeDetectionStrategy, Component, EventEmitter, Input, Output } from '@angular/core';
-import { LabelSetKind, LabelValue } from '@nuclia/core';
+import { LabelSetKind, Classification } from '@nuclia/core';
 import { map } from 'rxjs';
 import { LabelsService } from '../../../services/labels.service';
 
@@ -11,24 +11,24 @@ import { LabelsService } from '../../../services/labels.service';
 })
 export class LabelFieldComponent {
   @Input()
-  set selection(value: LabelValue[]) {
+  set selection(value: Classification[]) {
     this._selection = value || [];
   }
   get selection() {
     return this._selection;
   }
-  private _selection: LabelValue[] = [];
+  private _selection: Classification[] = [];
 
   @Input() kind: LabelSetKind = LabelSetKind.RESOURCES;
 
-  @Output() selectionChange = new EventEmitter<LabelValue[]>();
+  @Output() selectionChange = new EventEmitter<Classification[]>();
 
   labelSets$ = this.labelsService.getLabelsByKind(this.kind);
   hasLabels = this.labelSets$.pipe(map((labels) => labels && Object.keys(labels).length > 0));
 
   constructor(private labelsService: LabelsService) {}
 
-  updateSelection($event: LabelValue[]) {
+  updateSelection($event: Classification[]) {
     this.selection = [...$event];
     this.selectionChange.emit(this.selection);
   }