+++ apps/dashboard/src/app/resources/resource-list.component.ts
@@ -18,7 +18,7 @@ import { debounceTime, filter, map, switchMap, takeUntil, tap, toArray } from 'r
 import { ActivatedRoute, Router } from '@angular/router';
 import {
   deDuplicateList,
-  LabelValue,
+  Classification,
   Resource,
   RESOURCE_STATUS,
   ResourceList,
@@ -42,7 +42,7 @@ interface KeyValue {
   value: string;
 }
 
-interface ColoredLabel extends LabelValue {
+interface ColoredLabel extends Classification {
   color: string;
 }
 
@@ -103,7 +103,7 @@ export class ResourceListComponent implements OnInit, OnDestroy {
     }),
   );
   labelSets$ = this.sdk.currentKb.pipe(switchMap((kb) => kb.getLabels()));
-  currentLabelList: LabelValue[] = [];
+  currentLabelList: Classification[] = [];
 
   constructor(
     private sdk: SDKService,
@@ -289,8 +289,9 @@ export class ResourceListComponent implements OnInit, OnDestroy {
             resource,
             labels: [],
           };
-          if (resource.usermetadata?.classifications) {
-            resourceWithLabels.labels = resource.usermetadata.classifications.map((label) => ({
+          const labels = resource.getClassifications();
+          if (labels.length > 0) {
+            resourceWithLabels.labels = labels.map((label) => ({
               ...label,
               color: labelSets[label.labelset]?.color || '#ffffff',
             }));
@@ -346,7 +347,7 @@ export class ResourceListComponent implements OnInit, OnDestroy {
     }
   }
 
-  updateLabelList($event: LabelValue[]) {
+  updateLabelList($event: Classification[]) {
     this.currentLabelList = $event;
   }
 
@@ -356,7 +357,7 @@ export class ResourceListComponent implements OnInit, OnDestroy {
         const updatedResource = {
           usermetadata: {
             ...resource.usermetadata,
-            classifications: this.mergeExistingAndSelectedLabels(resource.usermetadata?.classifications),
+            classifications: this.mergeExistingAndSelectedLabels(resource.getClassifications()),
           },
         };
         return resource.modify(updatedResource).pipe(
@@ -385,7 +386,7 @@ export class ResourceListComponent implements OnInit, OnDestroy {
     }
   }
 
-  private mergeExistingAndSelectedLabels(classifications: LabelValue[] | undefined): LabelValue[] {
+  private mergeExistingAndSelectedLabels(classifications: Classification[] | undefined): Classification[] {
     if (!classifications) {
       return this.currentLabelList;
     }